services/audio/requirements.txt
openwakeword
pyaudio
numpy
flask
paho-mqtt
